Feat : Minor code reactors.

+++ shared/src/commonMain/kotlin/com/digitalcrafts/kmmweather/shared/models/internal/LatLong.kt
@@ -3,10 +3,10 @@ package com.digitalcrafts.kmmweather.shared.models.internal
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class LatLong(val lat: Long, val long: Long) {
+data class LatLong(val lat: Double, val long: Double) {
 
     companion object {
-        fun LatLong.toLatLon(): Pair<Long, Long> = Pair(this.lat, this.long)
-        fun Pair<Long, Long>.fromLatLon(): LatLong = LatLong(this.first, this.second)
+        fun LatLong.toLatLon(): Pair<Double, Double> = Pair(this.lat, this.long)
+        fun Pair<Double, Double>.fromLatLon(): LatLong = LatLong(this.first, this.second)
     }
 }
